Extract term attributes parsing
Parsing TERMATTRIBUTE column requires quite a lot of logic.  It makes
sense to extract it to a brand new class, especially that it is very
buggy and needs to be covered with tests.
